# openapi.yaml
openapi: 3.1.0
info:
  title: Writers Platform API
  description: A modern writing platform API built with FastAPI
  version: 1.0.0
paths:
...
  /api/v1/articles/:
    post:
      tags:
        - articles
      summary: Create Article
      description: |-
        Create a new article.

        Rate limit: 10 articles per minute per IP
        Requires: Valid JWT token, active user
      operationId: create_article_api_v1_articles__post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - articles
      summary: Get Articles
      description: |-
        Get articles with filtering, sorting, and pagination.

        Rate limit: 30 requests per minute per IP
        Authentication: Optional
      operationId: get_articles_api_v1_articles__get
      security:
        - HTTPBearer: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/ArticleStatus'
              - type: 'null'
            title: Status
        - name: category_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            title: Category Id
        - name: collection_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            title: Collection Id
        - name: author_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            title: Author Id
        - name: is_featured
          in: query
          required: false
          schema:
            anyOf:
              - type: boolean
              - type: 'null'
            title: Is Featured
        - name: tag
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Tag
        - name: search
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Search
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            title: Skip
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Limit
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
            default: created_at
            title: Sort By
        - name: sort_order
          in: query
          required: false
          schema:
            type: string
            pattern: ^(asc|desc)$
            default: desc
            title: Sort Order
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedArticleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/articles/{article_id}:
    get:
      tags:
        - articles
      summary: Get Article
      description: |-
        Get a specific article by ID.

        Rate limit: 60 requests per minute per IP
        Authentication: Optional (required for draft articles)
      operationId: get_article_api_v1_articles__article_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: article_id
          in: path
          required: true
          schema:
            type: integer
            title: Article Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
        - articles
      summary: Update Article
      description: |-
        Update an existing article.

        Rate limit: 20 updates per minute per IP
        Requires: Valid JWT token, article ownership or admin
      operationId: update_article_api_v1_articles__article_id__put
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: article_id
          in: path
          required: true
          schema:
            type: integer
            title: Article Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - articles
      summary: Delete Article
      description: |-
        Delete an article.

        Rate limit: 10 deletions per minute per IP
        Requires: Valid JWT token, article ownership or admin
      operationId: delete_article_api_v1_articles__article_id__delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: article_id
          in: path
          required: true
          schema:
            type: integer
            title: Article Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/articles/slug/{slug}:
    get:
      tags:
        - articles
      summary: Get Article By Slug
      description: |-
        Get a specific article by slug.

        Rate limit: 60 requests per minute per IP
        Authentication: Optional (required for draft articles)
      operationId: get_article_by_slug_api_v1_articles_slug__slug__get
      security:
        - HTTPBearer: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            title: Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/articles/{article_id}/status:
    patch:
      tags:
        - articles
      summary: Update Article Status
      description: |-
        Update article status (publish/unpublish/archive).

        Rate limit: 30 updates per minute per IP
        Requires: Valid JWT token, article ownership or admin
      operationId: update_article_status_api_v1_articles__article_id__status_patch
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: article_id
          in: path
          required: true
          schema:
            type: integer
            title: Article Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleStatusUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/articles/user/{user_id}:
    get:
      tags:
        - articles
      summary: Get User Articles
      description: |-
        Get articles by a specific user.

        Rate limit: 30 requests per minute per IP
        Authentication: Optional (required to see drafts)
      operationId: get_user_articles_api_v1_articles_user__user_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            title: User Id
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/ArticleStatus'
              - type: 'null'
            title: Status
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            title: Skip
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedArticleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/articles/{article_id}/publish:
    post:
      tags:
        - articles
      summary: Publish Article
      description: |-
        Publish a draft article.

        Rate limit: 20 publishes per minute per IP
        Requires: Valid JWT token, article ownership or admin
      operationId: publish_article_api_v1_articles__article_id__publish_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: article_id
          in: path
          required: true
          schema:
            type: integer
            title: Article Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/articles/{article_id}/unpublish:
    post:
      tags:
        - articles
      summary: Unpublish Article
      description: |-
        Unpublish a published article (move to draft).

        Rate limit: 20 unpublishes per minute per IP
        Requires: Valid JWT token, article ownership or admin
      operationId: unpublish_article_api_v1_articles__article_id__unpublish_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: article_id
          in: path
          required: true
          schema:
            type: integer
            title: Article Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ArticleCreate:
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
          description: Article title
        summary:
          anyOf:
            - type: string
              maxLength: 500
            - type: 'null'
          title: Summary
          description: Article summary
        content:
          type: string
          minLength: 1
          title: Content
          description: Article content
        meta_description:
          anyOf:
            - type: string
              maxLength: 160
            - type: 'null'
          title: Meta Description
          description: SEO meta description
        meta_keywords:
          anyOf:
            - type: string
              maxLength: 255
            - type: 'null'
          title: Meta Keywords
          description: SEO keywords
        category_id:
          anyOf:
            - type: integer
              exclusiveMinimum: 0
            - type: 'null'
          title: Category Id
          description: Category ID
        collection_id:
          anyOf:
            - type: integer
              exclusiveMinimum: 0
            - type: 'null'
          title: Collection Id
          description: Collection ID
        order_in_collection:
          anyOf:
            - type: integer
              exclusiveMinimum: 0
            - type: 'null'
          title: Order In Collection
          description: Order in collection
        allow_comments:
          type: boolean
          title: Allow Comments
          description: Allow comments on article
          default: true
        is_featured:
          type: boolean
          title: Is Featured
          description: Featured article
          default: false
        status:
          $ref: '#/components/schemas/ArticleStatus'
          description: Article status
          default: draft
        scheduled_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Scheduled At
          description: Scheduled publication time
        tag_names:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Tag Names
          description: List of tag names
      type: object
      required:
        - title
        - content
      title: ArticleCreate
      description: Schema for creating a new article.
...
    ArticleStatus:
      type: string
      enum:
        - draft
        - published
        - archived
        - scheduled
      title: ArticleStatus
      description: Article publication status.
    ArticleStatusUpdate:
      properties:
        status:
          $ref: '#/components/schemas/ArticleStatus'
          description: New article status
        scheduled_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Scheduled At
          description: Scheduled publication time
      type: object
      required:
        - status
      title: ArticleStatusUpdate
      description: Schema for updating article status only.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
    HTTPBearer:
      type: http
      scheme: bearer